organize, add golangci-lint, nolint a few basic warnings, separate the main api_test.go into tests/e2e and run with --tags e2e
